+++ algorithm-join-point-2/src/main/java/ru/vood/joinpoint2/infrastructure/flow/ActivityJoinPointOrderRunDaoService.kt
@@ -4,11 +4,11 @@ import ru.vood.joinpoint2.infrastructure.flow.data.JoinPointData
 
 interface ActivityJoinPointOrderRunDaoService {
 
-    fun nextJoinPoints(id: Int, joinPoint: String, flowType: String): Map<String, JoinPointData>
+    fun nextJoinPoints(id: Long, joinPoint: String, flowType: String): Map<String, JoinPointData>
 
-    fun prevJoinPoints(id: Int, joinPoint: String, flowType: String): Map<String, JoinPointData>
+    fun prevJoinPoints(id: Long, joinPoint: String, flowType: String): Map<String, JoinPointData>
 
-    fun getFirstJoinPoint(id: Int, flowType: String): Map<String, JoinPointData>
+    fun getFirstJoinPoint(id: Long, flowType: String): Map<String, JoinPointData>
 
-    fun getJoinPoint(id: Int, joinPoint: String): JoinPointData
+    fun getJoinPoint(id: Long, joinPoint: String): JoinPointData
 }
